various Manager updates and refactors

+++ Inventory/Item.cs
@@ -26,7 +26,7 @@ public class Item : MonoBehaviour
     [NonSerialized] public Sprite IconSprite;
     public string ColourHex { get; private set; } = "#FFFFFF";
     [NonSerialized] public Color ItemColor = Color.white;
-    public float Weight { get; private set; } = 1.0f;
+    public int Weight { get; private set; } = 1;
     public int Model { get; private set; } = 0;
     public int Bonus1 { get; private set; } = 0;
     public int Bonus2 { get; private set; } = 0;
